+++ example/lib/shared/decoration/spikes.dart
@@ -3,7 +3,7 @@ import 'dart:async' as async;
 import 'package:bonfire/bonfire.dart';
 import 'package:bonfire/util/priority_layer.dart';
 import 'package:example/manual_map/dungeon_map.dart';
-import 'package:example/util/common_sprite_sheet.dart';
+import 'package:example/shared/util/common_sprite_sheet.dart';
 
 class Spikes extends GameDecoration with Sensor {
   async.Timer? timer;