+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.9.1
+- Correction of loading visible collisions  on the map.
+- Adds a simple example.
+
+
 # 1.9.0
 - Update flame to `1.0.0-releasecandidate.15`
 - Adds  methods `changeJoystickTarget` in BonfireGame to make it easy to switch the default joystick events watcher.